# chess-backend/docs/api-contract/openapi.yaml
openapi: "3.0.3"
info:
  title: Chess99 API
  description: |
    REST API for the Chess99 real-time multiplayer chess platform.

    Covers authentication, game management, ratings, tournaments,
    tutorials, social features, WebSocket lifecycle endpoints, and
    push-notification device registration.

    All authenticated endpoints require a Bearer token obtained from
    the `/auth/login`, `/auth/register`, or OAuth endpoints.
    Tokens are Laravel Sanctum personal-access tokens.
  version: "1.0.0"
  contact:
    name: Chess99 Engineering
    url: https://chess99.com
  license:
    name: Proprietary
servers:
  - url: https://api.chess99.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local development
...
security: []
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
      description: Laravel Sanctum personal-access token

  # ── Reusable Parameters ───────────────────────────────────────────────────
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Magnus"
        email:
          type: string
          format: email
          example: "magnus@chess99.com"
        avatar_url:
          type: string
          format: uri
          nullable: true
          example: "https://api.chess99.com/storage/avatars/1_1700000000.jpg"
        rating:
          type: integer
          example: 1500
        is_provisional:
          type: boolean
          example: true
        games_played:
          type: integer
          example: 12
        peak_rating:
          type: integer
          example: 1520
        tutorial_xp:
          type: integer
          example: 350
        tutorial_level:
          type: integer
          example: 3
        current_skill_tier:
          type: string
          enum: [beginner, intermediate, advanced]
          nullable: true
          example: "beginner"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: "The given data was invalid."
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ["The email field is required."]
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Unauthenticated."
        status:
          type: string
          example: "error"
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            message: "Unauthenticated."
...
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            message: "Too Many Attempts."
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the rate limit resets
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
...
paths:
...
  /auth/google/mobile:
    post:
      operationId: googleMobileLogin
      tags: [Auth]
      summary: Google Sign-In for mobile apps
      description: |
        Accepts the `idToken` returned by the Google Sign-In SDK on
        Android/iOS. Creates an account on first login.
        Rate limited to 10 requests per minute.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idToken]
              properties:
                idToken:
                  type: string
                  description: Google ID token from mobile SDK
      responses:
        "200":
          description: Authenticated via Google
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/apple/mobile:
    post:
      operationId: appleMobileLogin
      tags: [Auth]
      summary: Sign in with Apple for mobile apps
      description: |
        Accepts the identity token from Apple Sign-In SDK.
        Creates an account on first login. Name and email are only
        provided on the first authorization.
        Rate limited to 10 requests per minute.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identity_token]
              properties:
                identity_token:
                  type: string
                authorization_code:
                  type: string
                  nullable: true
                user_name:
                  type: string
                  nullable: true
                  description: Only available on first authorization
                user_email:
                  type: string
                  format: email
                  nullable: true
                  description: Only available on first authorization
      responses:
        "200":
          description: Authenticated via Apple
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [Auth]
      summary: Refresh authentication token
      description: |
        Revokes the current token and issues a new one.
        Rate limited to 10 requests per minute.
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                device_name:
                  type: string
                  nullable: true
                  description: Optional label for the new token
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/revoke-all:
    post:
      operationId: revokeAllTokens
      tags: [Auth]
      summary: Revoke all tokens for current user
      description: Signs the user out of all devices.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: All tokens revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                  tokens_revoked:
                    type: integer
                    example: 3
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  # ── User / Profile ───────────────────────────────────────────────────────
...